+++ research/train_model.py
@@ -2,7 +2,7 @@ import multiprocessing
 from gensim.models import Word2Vec
 from gensim.models.word2vec import LineSentence
 
-from research import get_config
+from research.base import get_config
 
 if __name__ == "__main__":
     config = get_config()
